backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apturl==0.5.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dalle2==1.0.11
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
entrypoints==0.4
fasteners==0.14.1
Flask==2.0.1
Flask-Cors==3.0.10
frozenlist==1.3.3
future==0.18.2
httplib2==0.20.2
idna==2.10
importlib-metadata==4.6.4
itsdangerous==2.1.0
jeepney==0.7.1
Jinja2==3.0.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
numpy==1.24.2
oauthlib==3.2.0
olefile==0.46
openai==0.27.2
opencv-python==4.7.0.72
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.4.0
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-unixsocket==0.2.0
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
ssh-import-id==5.11
systemd-python==234
tomli==2.0.1
tqdm==4.65.0
ttystatus==0.38
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.6
usb-creator==0.3.7
wadllib==1.3.6
Werkzeug==2.0.2
xdg==5
xkit==0.0.0
yarl==1.8.2
zipp==1.0.0
